De-dup module registration
This simplifies adding a new Memory-only Firestore build
